+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+0.0.6 October 11, 2018
+  - Add Support for Inventory API : getBulkInventoryItem
+  - Add examples in README.md
+
 0.0.5 October 3, 2018
   - Add Support for Offer API : createOffer, updateOffer,getOffer, getOffers, publishOffer, deleteOffer, getListingFees, withdrawOffer
   - Add Support for Seller Location API : createInventoryLocation, updateInventoryLocation, enableInventoryLocation, disableInventoryLocation, getInventoryLocation, deleteInventoryLocation, getInventoryLocations